requirements.txt
flask
python-dotenv
googlesearch-python
google-generativeai
requests
numpy
beautifulsoup4 